[profile-handler] update to comply with data-driven requests
Currently, nukr.handlers.profile-handler namespace has an API which
complies with destructured requests, somewhat similar to how compojure
organizes and deals with them.

This has some limitations, because it doesn't allow bidirectional
matching of routes, which is a desirable for the project. Since this
feature is normally attached to data-driven requests, it's been proved
that refactoring the namespace so it complies with this paradigm.