[profile] fix wrong usage of contains? function
The current use of contains? function is not right, since it is not
meant to test wether a particular item is within a collection, but
rather to check is a collection contains a given key.

To solve this, substitute contains? by Java method .contains, which has
the same behaviour for collections.