[profile-handler] add spec to create profile request
Until now, there was not a way to test whether a particular request
contained valid data or not, when trying to create a profile.

This commit creates a spec to define what can be considered a good
request, in such a way that becomes possible to return HTTP status 400
for those who doesn't meet these requirements.