[suggestion] create suggestions entity
Since the source code of functions related to suggestions doesn't have
much in common with the remaining profile namespace code, it seems to be
more productive to separate it into a different file.

This makes the code more logically organized and delegates better what part
of the code is in charge of a given feature.