[system] init-system! can bind to given port or default
The current behaviour had a hardcoded port into the system namespace,
which always bound the server to 4000 port. Given that this is quite
limiting, the new structure allows for users to decided in which port
their aleph server will be bind by providing it as an argument to `lein
run` command.