[server] substitute aleph by jetty as server adapter
Aleph has some serious issues. It can't be tested properly because of an
unpredictable issue, which causes an exception to be raised when the
server is closed.

Substituting aleph by jetty as the server adapter put us on control of
the testing once again. This not only allows testing, but also prevents
any possible leak due to servers left running during the tests.